added delete, update, getById endpoints, changed int to Long id

+++ src/main/java/com/jw/dto/OrderResponse.java
@@ -3,4 +3,4 @@ package com.jw.dto;
 import lombok.Builder;
 
 @Builder
-public record OrderResponse(int id, String name) {}
+public record OrderResponse(Long id, String name) {}
